Remove DatasetManager dependency in ImageFile
DatasetManager was only used to retrieve .config and had no other purpose. Now ImageFile reads .config by itself as a class variable which is loaded when the class ImageFile is imported. Previously it was an instance variable copied from the dataset manager.

This allows us to pickle ImageFile objects, and also makes code cleaner and more stable.